// apps/desktop/src/ipc/generated/PlanState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlanApprover } from "./PlanApprover";

export type PlanState = { "type": "proposed" } | { "type": "approved", by: PlanApprover, } | { "type": "rejected", message: string | null, } | { "type": "superseded" };